Bug 984365 - Refactor 'addon' commands; r=mratcliffe
Remove the initial cache step, and replace it with a new 'addon' type.
The cache was used to generate {name:..., value:...} objects for the
selection type. Previously the value was set to addon.name, but now
we are setting it to just addon. This greatly simplifies the enable and
disable commands because they already have an addon and don't need to
look one up.