Bug 1013897 - Allow generation of more than one report for a target; r=harth
Previously there was one box which contained the template, so the second
time we created a report, the template was used. Now there are 2 boxes
the first containing the template which we clone before altering and
moving it to the second box.

At the same time we fix the indentation of the existing style editor
components, and use a generator function to reduce indentation.