Use a single parameter for the "error" event.
Having multiple parameters was getting confusing because sometimes it
wasn't clear what was the type of error and what were the parameters
and the list was getting longer and more confusing.

Also multiple parameters interact badly with promises where there can
only be one value, so using the the promise return from once() breaks.