Include imgur support
We've had this attachment [1] to bug 992386, outstanding for ages. I'm
just totally breaking it, to the point where if we don't include it now
then it will rot forever, so I'm including it now.

This is a basic port of something that was close to r+, so my hunch is
that this changeset doesn't need a significant review. I'm checking this
is now so you can see the steps.

That said, there are some things that I know I've broken here, so you
might want to check ahead to the end result.

It's also possible that by the end this will probably look a lot clearer
viewed from scratch.

[1]: https://bugzilla.mozilla.org/page.cgi?id=splinter.html&bug=992386&attachment=8507127