Complete pass 1 of tidying up the screenshot command
But we're not done yet. Currently neither the clipboard option nor the
imgur option works, this could have something to do with running in a
content process, and perhaps with sandboxing even if that's not a
problem now, then it will be in the future, so maybe we need to move
that to the converter.

Also the --chrome option is probably not going to work in the content
process, so since commands can only run in a single process, I think
we're going to need some sort of a 'chromescreenshot' command, which can
probably share almost all the code with the 'screenshot' command, but
needs a different document to work from.