Fix where a number of commands are run
Most non-trivial commands are only going to run correctly on the server or
the client, and our tests are generally good at flushing this out.

Note on the 'jsb' command: I'm not sure this is a particularly useful command,
I'm half of the mind that we should remove it, but I think that's another job.
However I did add a bit of detail to the exec test.