Fix 'edit' command
This is a tricky one. The 'edit' command uses the toolbox, so it's
clearly runAt:client, but it uses the 'resource' type which accesses the
DOM.

TL;DR: (Option 3)

The solutions that I've looked at are:

1. Create a new clientResource type that works for commands that
runAt:client. I'm sure there are Debugger/StyleEditor, etc APIs to do
the heavy work, but I'm hoping there is a quicker cheat.

2. Create a hidden command that's runAt:server with a resource
parameter, and then use the 'remote' type to point the real edit command
to use the resource type of the hidden command.

    exports.items = [
      {
        item: "command",
        runAt: "server",
        name: "edit",
        params: [
          {
            name: "resource",
            type: {
              name: "remote",
              commandName: "__hidden_edit",
              paramName: "resource",
            },
          }
        ],
        exec: function(args, context) {
          // ...
        }
      },
      {
        item: "command",
        runAt: "client",
        name: "__hidden_edit",
        hidden: true,
        params: [{
          name: "resource",
          type: { name: "resource", include: "text/css" },
        }],
        exec: () => { }
      }
    ];

This is neat, but the remote type expects that the 'import' process (as
defined by system.connectFront() will inject the front. Since there is
no import process here, that won't work. Maybe there is a non-hacky way
to get at the front?

3. Have the edit command be runAt:client and simply return 'args' (or at
least a JSONable version of that) and have a converter that actually
opens the style editor. In some ways you can see the style editor as the
'view' anyway, but I don't think this is ideal, however it is simple so
it's what I've done for now.

Also, the test suite was testing values, which it can't do for remote
types.