Fix for "Exception on toolbar open with text"
STR:
1. Type something into the developer toolbar
2. Close the toolbar
3. Open the toolbar

Actual: the toolbar doesn't open in step 3.

The completer updates assuming the text in the input element, but the
requisition doesn't know about that yet, so they get confused.

We need to update the requisition before the completer gets involved,
so we call update (which is asynchronous).

This means lots of text gets indented, but this is really just a one
line change.